Added NewType and moved Circuit
