+++ machq/types/__init__.py
@@ -1,1 +1,2 @@
-from machq.types._stim_types import Circuit, Qubit
+from machq.types._basic_types import Qubit, Distance
+from machq.types._circuit_ import Circuit
